move to MessagePack.Unity.Mathmatics.asmdef
