+++ src/MessagePack.UnityClient/Assets/Scripts/MessagePack/Mathmatics.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 73eac4caceef18a4ca9aae640f0cd3dd
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
