+++ src/MessagePack.UnityClient/Assets/Scripts/MessagePack/Mathmatics/MessagePack.Unity.Mathmatics.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f11e9f7ce55e6c74a8a34f51e1c1678f
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
