+++ src/MessagePack.UnityClient/Assets/Scripts/MessagePack/Mathmatics/UnityMathmaticsFormatters.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bc412b4c5b74fd040b0ad898b1a5af49
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
