+++ src/MessagePack.UnityClient/Assets/Scripts/MessagePack/Mathmatics/UnityMathmaticsResolver.cs.meta
